pyproject.toml
[tool.poetry]
name = "haloart"
version = "0.1.0"
description = "shoping list"
authors = ["Alexey Inkov <alexo982@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
django-filter = "^23.5"

[tool.poetry.group.dev.dependencies]
black = "23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
